Add files via upload
 Event Management System :

Using the application, users are able to choose a theme, set the guest count and select a photography offer. It uses a detailed layout so you can book, review wedding themes, choose your meals and decide on photography. It figures out the overall expense and generates a summary into a text file.

Highlights:

Quick and easy form for reserving events

There are 4 theme areas to choose from.

Works out the expenses for food, decorations and taking pictures

Stores the event summary in a .txt file

Looks to see that the phone, email and given date are correct