Add files via upload
Periodic Table Explorer :

Using Java Swing, this application lets you discover the elements found in the periodic table. It is possible to look for elements by referring to their number, group, period or block. The information about symbol, weight and valence can be found easily in the interface.

Highlights:

You can search for elements by their group, when they were formed, what block they are in or by their number.

Information about 118 elements are provided.

Simple and straightforward way to interact with the device

All insight is presented in a scrollable text box.

Created with Swing and filters presented in a dropdown menu